backend/requirements.txt
Django==4.0.5
django-cors-headers==3.11.0
djangorestframework==3.13.1
psycopg2==2.9.3
python-decouple==3.6